Fusion de UserRdvWizard dans RdvBookingForm
- Création de RdvBuilderConcern pour la logique partagée de construction de RDV
- RdvBookingForm inclut maintenant directement le concern et gère tout le flux usager
- PrescripteurRdvWizard utilise le concern au lieu d'hériter de UserRdvWizard
- Suppression de UserRdvWizard devenu obsolète
