+++ spec/controllers/users/rdv_wizard_steps_controller_spec.rb
@@ -7,21 +7,22 @@ RSpec.describe Users::RdvWizardStepsController, type: :controller do
     let(:starts_at) { Time.zone.parse("2020-03-03 10h00") }
     let!(:mock_creneau) { instance_double(Creneau) }
     let!(:mock_rdv) { build(:rdv, starts_at: starts_at, users: [user], created_by: user) } # cannot use instance_double because it breaks pundit inference
-    let(:mock_user_rdv_wizard) { instance_double(UserRdvWizard, creneau: mock_creneau, rdv: mock_rdv) }
+    let(:mock_rdv_booking_form) { instance_double(Users::RdvBookingForm, creneau: mock_creneau, rdv: mock_rdv) }
 
     before { travel_to Date.parse("2020-03-01").in_time_zone + 8.hours }
 
     context "logged in user" do
       before do
-        allow(UserRdvWizard).to \
+        allow(Users::RdvBookingForm).to \
           receive(:new).with(
-            user,
-            hash_including(
+            user: user,
+            attributes: hash_including(
               "motif_id" => motif.id.to_s,
               "lieu_id" => lieu.id.to_s,
               "starts_at" => starts_at.to_s
-            )
-          ).and_return(mock_user_rdv_wizard)
+            ),
+            domain: anything
+          ).and_return(mock_rdv_booking_form)
       end
 
       context "when signed in" do
@@ -44,14 +45,11 @@ RSpec.describe Users::RdvWizardStepsController, type: :controller do
     end
 
     context "usager connecté via une invitation" do
-      let(:mock_step1_wizard) { instance_double(UserRdvWizard::Step1, creneau: mock_creneau, rdv: mock_rdv) }
-
       before do
         sign_in user
         user.signed_in_with_invitation_token!
         allow(controller).to receive(:current_user).and_return(user)
-        allow(UserRdvWizard::Step1).to receive(:new).and_return(mock_step1_wizard)
-        allow(Users::RdvBookingForm).to receive(:new).and_return(instance_double(Users::RdvBookingForm))
+        allow(Users::RdvBookingForm).to receive(:new).and_return(mock_rdv_booking_form)
       end
 
       it "le step 1 pointe vers le step 3 comme prochaine étape" do
@@ -62,12 +60,10 @@ RSpec.describe Users::RdvWizardStepsController, type: :controller do
 
     context "usager connecté via ProConnect" do
       let!(:user) { create(:user, pro_connect_openid_sub: "some-openid-sub") }
-      let(:mock_step1_wizard) { instance_double(UserRdvWizard::Step1, creneau: mock_creneau, rdv: mock_rdv) }
 
       before do
         sign_in user
-        allow(UserRdvWizard::Step1).to receive(:new).and_return(mock_step1_wizard)
-        allow(Users::RdvBookingForm).to receive(:new).and_return(instance_double(Users::RdvBookingForm))
+        allow(Users::RdvBookingForm).to receive(:new).and_return(mock_rdv_booking_form)
       end
 
       it "le step 1 pointe vers le step 3 comme prochaine étape" do
